+++ views/chart.py
@@ -1,0 +1,7 @@
+from mesa.visualization.modules import ChartModule
+
+
+chart = ChartModule(
+    [{"Label": "Altruistas", "Color": "Black"}],
+    data_collector_name="datacollector",
+)
